+++ docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Gopjrt Changelog
 
+<<<<<<< main
+=======
+# Next
+
+* Moved `stablehlo` to its own repository.
+* Updated dependencies.
+* Package `dtypes`:
+    * Added `DType.Bits` field indicating the number of bits in the data type -- supporting sub-byte types including
+      quantized types.
+
+>>>>>>> local
 # v0.8.0 - 2025/08/20 Convolution API changes
 
 * **API change to convolutions**: